backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aidin <aidinror@gmail.com>", "Nikita <20nik.nosov21@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
syncpg = "^1.1.2"
pydantic = {extras = ["email"], version = "^2.9.0"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"
greenlet = "^3.0.3"


[tool.poetry.group.dev.dependencies]
setuptools = "^74.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
